pyproject.toml
[tool.poetry]
name = "gol"
version = "0.1.2"
description = "Game of life python demonstration project"
readme = "README.md"
authors = ["Jeroen Arbon <jeroen.arbon@gmail.com>"]
packages = [
    { include = "gol", from = "src" },
]
classifiers = [
        "Intended Audience :: Developers",
        "License :: OSI Approved :: MIT License",
        "Topic :: Games/Entertainment :: Simulation",
        "Programming Language :: Python :: 3.5",
        "Programming Language :: Python :: 3.6",
        "Programming Language :: Python :: 3.7",
        "Programming Language :: Python :: Implementation :: PyPy"
]

[tool.poetry.dependencies]
python = "^3.6"
click = "^7.0"
numpy = "^1.16"
python-decouple = "^3.1"
colorama = "^0.4.1"

black = {version = "^18.3-alpha.0", allows-prereleases = true, optional = true}
flake8 = {version = "^3.7", optional = true}

[tool.poetry.dev-dependencies]
tox = "^3.13"
coverage = "^4.5"
pytest = "^5.0"

[tool.poetry.extras]
linters = ["black", "flake8"]

[tool.poetry.scripts]
gol = "gol.start:game_of_life_command"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
